first view use case works

+++ src/main/scala/com/gnosly/fluentsequence/view/fixedwidth/Util.scala
@@ -1,7 +1,7 @@
 package com.gnosly.fluentsequence.view.fixedwidth
 
 object Util {
-	def middle(name: String, padding: Int): String = {
+	def middle(name: String, padding: Long): String = {
 		if (padding <= 1) {
 			return name
 		}
@@ -9,6 +9,6 @@ object Util {
 		" " + middle(name, padding - 2) + " "
 	}
 
-	def r(pattern: String, count: Int): String =
-		(0 until count).map(_ => pattern).reduce(_ + _)
+	def r(pattern: String, count: Long): String =
+		(0 until count.toInt).map(_ => pattern).reduce(_ + _)
 }
